--- MusicDisk1/src/annotations/FindByName.java
@@ -1,8 +1,0 @@
-package annotations;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface FindByName {
-}
