--- MusicDisk1/tracks.txt
@@ -1,37 +1,0 @@
-Robbie Williams;Angels;Life thru a Lens;1997;Pop-rock;4:24;4;
-Green Day;American Idiot;American Idiot;2004;Punk;2:54;7;
-Green Day;Jesus of Suburbia;American Idiot;2004;Punk;9:08;15;
-Green Day;Holiday;American Idiot;2004;Punk;3:52;5;
-Green Day;Boulevard of Broken Dreams;American Idiot;2004;Punk;4:20;6;
-Green Day;Are We the Waiting;American Idiot;2004;Punk;2:42;3;
-Green Day;St. Jimmy;American Idiot;2004;Punk;2:56;7.8;
-Green Day;Give Me Novacaine;American Idiot;2004;Punk;3:25;7;
-Green Day;Letterbomb;American Idiot;2004;Punk;4:05;9;
-Green Day;Wake Me Up When September Ends;American Idiot;2004;Punk;4:45;11;
-Green Day;Homecoming;American Idiot;2004;Punk;9:18;15.5;
-Green Day;Whatsername;American Idiot;2004;Punk;4:14;10;
-Nine Inch Nails;Mr. Self Destruct;The Downward Spiral;1994;Industrial;4:30;11;
-Nine Inch Nails;Piggy;The Downward Spiral;1994;Industrial;4:24;10;
-Nine Inch Nails;Heresy;The Downward Spiral;1994;Industrial;3:54;8;
-Nine Inch Nails;March of the Pigs;The Downward Spiral;1994;Industrial;2:58;7;
-Nine Inch Nails;Closer;The Downward Spiral;1994;Industrial;6:13;14;
-Nine Inch Nails;Ruiner;The Downward Spiral;1994;Industrial;4:58;9;
-Nine Inch Nails;The Becoming;The Downward Spiral;1994;Industrial;5:31;13;
-Nine Inch Nails;I Do Not Want This;The Downward Spiral;1994;Industrial;5:41;15;
-Nine Inch Nails;Big Man with a Gun;The Downward Spiral;1994;Industrial;1:36;8;
-Nine Inch Nails;A Warm Place;The Downward Spiral;1994;Industrial;3:22;17;
-Nine Inch Nails;Eraser;The Downward Spiral;1994;Industrial;4:54;16;
-Nine Inch Nails;Reptile;The Downward Spiral;1994;Industrial;6:51;13;
-Nine Inch Nails;The Downward Spiral;The Downward Spiral;1994;Industrial;3:57;9.2;
-Nine Inch Nails;Hurt;The Downward Spiral;1994;Industrial;6:13;17;
-David Bowie;Changes;Hunky Dory;1971;Art Rock;3:37;5;
-David Bowie;Oh! You Pretty Things;Hunky Dory;1971;Art Rock;3:12;4.8;
-David Bowie;Eight Line Poem;Hunky Dory;1971;Art Rock;2:55;4.3;
-David Bowie;Life on Mars?;Hunky Dory;1971;Art Rock;3:53;6;
-David Bowie;Kooks;Hunky Dory;1971;Art Rock;2:53;3.7;
-David Bowie;Quicksand;Hunky Dory;1971;Art Rock;5:08;7;
-David Bowie;Fill Your Heart;Hunky Dory;1971;Art Rock;3:07;8;
-David Bowie;Andy Warhol;Hunky Dory;1971;Art Rock;3:56;8.4;
-David Bowie;Song for Bob Dylan;Hunky Dory;1971;Art Rock;4:12;9;
-David Bowie;Queen Bitch;Hunky Dory;1971;Art Rock;3:18;8;
-David Bowie;The Bewlay Brothers;Hunky Dory;1971;Art Rock;5:22;10
