+++ src/annotations/FindBySize.java
@@ -1,0 +1,8 @@
+package annotations;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface FindBySize {
+}
