+++ src/exceptions/DiskTypeException.java
@@ -1,0 +1,11 @@
+package exceptions;
+
+/**
+ * Exception is thrown when disk check fails
+ */
+public class DiskTypeException extends Exception {
+
+    public DiskTypeException(String message){
+        super(message);
+    }
+}
