+++ src/interfaces/IParameters.java
@@ -1,0 +1,15 @@
+package interfaces;
+
+/**
+ * Interface for common parameters in Track and Disk classes.
+ */
+public interface IParameters {
+
+    String getName();
+
+    double getSize();
+
+    void setName(String name);
+
+    void trackInfo();
+}
